+++ libfuturize/fixes/__init__.py
@@ -17,7 +17,8 @@ lib2to3_fix_names_stage1 = set([
     'lib2to3.fixes.fix_isinstance',
     'lib2to3.fixes.fix_methodattrs',
     'lib2to3.fixes.fix_ne',
-    'lib2to3.fixes.fix_next',
+    # 'lib2to3.fixes.fix_next',         # would replace ``next`` method names
+                                        # with ``__next__``.
     'lib2to3.fixes.fix_numliterals',    # turns 1L into 1, 0755 into 0o755
     'lib2to3.fixes.fix_paren',
     # 'lib2to3.fixes.fix_print',        # see the libfuturize fixer that also
@@ -57,6 +58,7 @@ lib2to3_fix_names_stage2 = set([
     'lib2to3.fixes.fix_long',
     'lib2to3.fixes.fix_map',
     # 'lib2to3.fixes.fix_metaclass', # causes SyntaxError in Py2! Use the one from ``six`` instead
+    'lib2to3.fixes.fix_next',
     'lib2to3.fixes.fix_nonzero',     # TODO: cause this to import ``object`` and/or add a decorator for mapping __bool__ to __nonzero__
     'lib2to3.fixes.fix_operator',    # we will need support for this by e.g. extending the Py2 operator module to provide those functions in Py3
     'lib2to3.fixes.fix_raw_input',
@@ -69,6 +71,9 @@ lib2to3_fix_names_stage2 = set([
 libfuturize_fix_names_stage1 = set([
     'libfuturize.fixes.fix_absolute_import',
     'libfuturize.fixes.fix_division',
+    'libfuturize.fixes.fix_next_call',  # obj.next() -> next(obj). Unlike
+                                        # lib2to3.fixes.fix_next, doesn't change
+                                        # the ``next`` method to ``__next__``.
     'libfuturize.fixes.fix_print_with_import',
     'libfuturize.fixes.fix_raise',
     'libfuturize.fixes.fix_order___future__imports',  # TODO: consolidate to a single line to simplify testing